test: fix Meteor app tests
They are now all running reliably, but
issues with the fixtures and the tests
need to be fixed
