+++ imports/meteor-app-tests/get-sitemap-xml.app-test.js
@@ -1,16 +1,17 @@
 import Random from "@reactioncommerce/random";
 import { expect } from "meteor/practicalmeteor:chai";
 import Reaction from "/imports/plugins/core/core/server/Reaction";
-import { Sitemaps } from "/imports/plugins/included/sitemap-generator/server/lib/collections/sitemaps";
+import { Sitemaps } from "/imports/plugins/included/sitemap-generator/lib/collections/sitemaps";
 import getSitemapXML from "/imports/plugins/included/sitemap-generator/server/lib/get-sitemap-xml";
 
-before((done) => {
-  Reaction.onAppStartupComplete(() => {
-    done();
-  });
-});
-
 describe("getSitemapXML", () => {
+  before(function (done) {
+    this.timeout(20000);
+    Reaction.onAppStartupComplete(() => {
+      done();
+    });
+  });
+
   it("should return an empty string if no sitemap by handle for given shop is found", () => {
     Sitemaps.remove({});
     const xml = getSitemapXML("FAKE_SHOP_ID", "FAKE_HANDLE");
