+++ imports/meteor-app-tests/shops.app-test.js
@@ -11,14 +11,15 @@ import Reaction from "/imports/plugins/core/core/server/Reaction";
 import ReactionError from "@reactioncommerce/reaction-error";
 import { Shops } from "/lib/collections";
 
-before((done) => {
-  Reaction.onAppStartupComplete(() => {
-    Fixtures();
-    done();
-  });
-});
-
 describe("Shop Methods", function () {
+  before(function (done) {
+    this.timeout(20000);
+    Reaction.onAppStartupComplete(() => {
+      Fixtures();
+      done();
+    });
+  });
+
   beforeEach(function () {
     return Shops.remove({});
   });
